LayoutTests/media/encrypted-media/mock-MediaKeySession-update.html

<!DOCTYPE html>
<html>
<head>
    <script src=../video-test.js></script>
    <script type="text/javascript">
    var mock;
    var promise;
    var mediaKeySystemAccess;
    var mediaKeys;
    var mediaKeySession;
    var capabilities = {};
    var kids;

    function doTest()
    {
        if (!window.internals) {
            failTest("Internals is required for this test.")
            return;
        }

        run('internals.initializeMockMediaSource()');
        run('mock = internals.registerMockCDM()');
        run('mock.supportedDataTypes = ["keyids"]');
        run('capabilities.initDataTypes = ["keyids"]');
        run(`capabilities.videoCapabilities = [{ contentType: 'video/mock; codecs="mock"' }] `);
        run('promise = navigator.requestMediaKeySystemAccess("org.webkit.mock", [capabilities])');
        shouldResolve(promise).then(gotMediaKeySystemAccess, failTest);
    }

    function next() {
        if (!tests.length) {
            mock.unregister();
            endTest()
            return;
        }

        var nextTest = tests.shift();
        consoleWrite('');
        nextTest();
    }

    function gotMediaKeySystemAccess(result) {
        mediaKeySystemAccess = result;
        next();
    }

    function gotMediaKeys(result) {
        mediaKeys = result;
        next();
    }

    function stringToUInt8Array(str)
    {
       var array = new Uint8Array(str.length);
       for (var i=0; i<str.length; i++)
            array[i] = str.charCodeAt(i);
       return array;
    }

    tests = [
        function() {
            run('promise = mediaKeySystemAccess.createMediaKeys()');
            shouldResolve(promise).then(gotMediaKeys, failTest);
        },

        function() {
            consoleWrite('Using a non-callable MediaKeySession should reject.');
            run('mediaKeySession = mediaKeys.createSession("temporary")');
            testExpected('typeof mediaKeySession', 'object');
            run('promise = mediaKeySession.update(stringToUInt8Array("invalid-state"))');
            shouldReject(promise).then(next, next);
        },

        function() {
            consoleWrite('Using a zero-length response should reject.');
            run('kids = JSON.stringify({ kids: [ "MTIzNDU=" ] })');
            run('mediaKeySession = mediaKeys.createSession("temporary")');
            run('promise = mediaKeySession.generateRequest("keyids", stringToUInt8Array(kids))');
            shouldResolve(promise).then(function() {
                run('promise = mediaKeySession.update(new Uint8Array(0))');
                shouldReject(promise).then(next, next);
            }, next);
        },

        function() {
            consoleWrite('Using a non-sanitizable response should reject.');
            run('kids = JSON.stringify({ kids: [ "MTIzNDU=" ] })');
            run('mediaKeySession = mediaKeys.createSession("temporary")');
            run('promise = mediaKeySession.generateRequest("keyids", stringToUInt8Array(kids))');
            shouldResolve(promise).then(function() {
                run('promise = mediaKeySession.update(stringToUInt8Array("invalid-response"))');
                shouldReject(promise).then(next, next);
            }, next);
        },

        function() {
            consoleWrite('Using a sanitizable response should resolve.');
            run('kids = JSON.stringify({ kids: [ "MTIzNDU=" ] })');
            run('mediaKeySession = mediaKeys.createSession("temporary")');
            run('promise = mediaKeySession.generateRequest("keyids", stringToUInt8Array(kids))');
            shouldResolve(promise).then(function() {
                run('promise = mediaKeySession.update(stringToUInt8Array("valid-response"))');
                shouldResolve(promise).then(next, next);
            }, next);
        },

        function() {
            consoleWrite('Using a sanitizable response with invalid format should resolve.');
            run('kids = JSON.stringify({ kids: [ "MTIzNDU=" ] })');
            run('mediaKeySession = mediaKeys.createSession("temporary")');
            run('promise = mediaKeySession.generateRequest("keyids", stringToUInt8Array(kids))');
            shouldResolve(promise).then(function() {
                run('promise = mediaKeySession.update(stringToUInt8Array("valid-response invalid-format"))');
                shouldReject(promise).then(next, next);
            }, next);
        },
    ];
    </script>
</head>
<body onload="doTest()">
</body>
</html>
